Add test for discarding of drafts when annotations are saved
The logic for handling this case in the controller is slightly
indirect. The draft is discarded not when the call to update
the annotation on the server succeeds but whenever the client
is notified that an annotation update has been committed,
signified via its 'updated' property changing.

This happens both for saves triggered locally and also
updates made in other H instances.