Remove old feature flag data on app boot
When a feature flag is removed from the codebase, it will remain in the
database. This could potentially cause very surprising issues in the
event that a feature flag with the same name (but a different meaning)
is added at some point in the future.

This commit adds a function which removes unknown feature flags from the
database, and configures it to run at application startup.