Speed up database tests using SAVEPOINT/ROLLBACK TO SAVEPOINT
Tests that touched the database were doing a full drop and recreation of
the entire schema on every test. This meant that even a no-op test took
an appreciable fraction of a second to run.

Slow tests mean sad developers, so this commit changes how we run such
tests: the database schema is dropped and recreated only once -- at the
start of the test session -- and test isolation is instead guaranteed
using PostgreSQL's SAVEPOINT feature.

For each test function, we run the following in the database:

1. BEGIN; -- Open a new transaction. This is implicit.
2. SAVEPOINT some_savepoint_name; -- Establish a savepoint within the
   current transaction. For more on savepoints, see [the PostgreSQL
   docs][1].
3. Run the test function.
4. ROLLBACK TO SAVEPOINT some_savepoint_name; -- Roll the state of the
   database back to that at the savepoint.

This results in database tests that are between 5 and 10x faster, and
our overall test suite runs in less than half the time.

[1]: http://www.postgresql.org/docs/current/static/sql-savepoint.html