Generalise H_BUILDEXT to H_SCRIPT
We need the ability to turn off particular parts of application startup
in the general case that a script is bootstrapping the application
rather than a WSGI server.

This commit renames the H_BUILDEXT environment variable to H_SCRIPT and
ensures that it's set in h/script.py as well as h/buildext.py.

In addition, we move the code to remove old feature flags into a
NewApplication event subscriber rather than a configuration action, as
it is not "configuration" and needs a fully-configured application to
run.