Don't copy scopes from parents to replies
We were copying scopes from parents to replies because we were searching
on the scope field (for example, to find all annotations of a given
document) and we needed replies to show up in those search results.

But the search code has now been changed to always search for top-level
annotations only (replies are always filtered out) and then always do a
second Elasticsearch query for all replies to all annotations that
matched the first query.

So we're no longer searching against the scope fields of replies,
and no longer need these fields.