Enable staff to look up user's email addresses
Allow staff to access the /admin index page and to use some of the admin pages
within (so far only the /admin/users page).

Assign a different permission to each admin page ('admin_features',
'admin_nipsa' etc) then on the root resource assign some of these permissions
to admins and some to staff (or to both admins and staff).

This enables staff users to access the admin index page and some of the admin
pages within.

Then tweak the admin template to only show links to the admin pages that the
user has permission to use.

The "admin" permission, previously assigned to admin users, is removed since
it's no longer used.