Enable highlights by default everywhere
Enable highlights by default in the Chrome extension, bookmarklet and embed.

They were already enabled by default in the Chrome extension and
bookmarklet so this only changes the embed.

Highlights are turned on by default in annotator/host.coffee if no
showHighlights option is found.

The extension and bookmarklet no longer need to set showHighlights: true as
this is now the default. The embed (which was not setting showHighlights: true
and so was getting the default setting of false) will now get the new default
setting of true.

For people embedding h into their own pages, they can turn off highlights by
default by including:

    <script>
      window.hypothesisConfig=function(){return{showHighlights:false}};
    </script>