Move the hypothes.is front page into the app
Initial move of the hypothes.is front page into the app, much tidying
up and migrating of resources stil to do.

For this commit I did a view-source on https://hypothes.is/ and
copy-pasted it into home.html.jinja2, then:

- Routed "/" to home.html.jinja2, replaces the previous "/" route that went to
  the help page
- Reformatted the HTML to make it more manageable
- Removed the link to the empty RSS comments feed
- Removed tags in the <head> to do with installing the site as a web app on
  iOS, Android and Windows Mobile
- Removed WordPress RSD EditURI metadata tags
- Removed wlwmanifest.xml metadata link
- Removed the <link rel="shortlink" whose value was the same as the
  canonical link
- Removed the Google Analytics JavaScript
- Removed the "You need to update your browser" message and browsehappy.com
  link that was shown to IE < 9 users
- Removed the broken <script> link to showhighlights.js
- Removed inline JS and CSS to do with "WordPress emojis"

This "works" but still pulls all of the resources (CSS, images, js...) from the
production WordPress instance. Much tidying up still to do.