Remove main.js
The remove the <script> link to the main.js file of the WordPress instance.
Replace this with some inline JavaScript that does the one thing that (as far
as I know) main.js was being used for on the front page: to show the
install-chrome-extension button in Chrome but show the install-bookmarklet
button in all other browsers.

Changing this to a tiny bit of inline js also means that the install-chrome-
extension button is visible immediately in Chrome, instead of seeing the
bookmarklet button for a second then seeing it change.