Fix the account links on the front page
Fix the Sign in/Sign out/Create an account links in the top right of the
front page.

When logged in you see:

* Your username, links to your My Annotations page
* Sign out, redirects to front page after logout

When not logged in you see:

* Sign in, links to /login and redirects to /stream after login
* Create an account, links to /register and redirects to front page

Remove the "You have logged out" flash message. Since these flash messages are
only shown in the Angular app, and logging out redirects you to the front page,
the message is not shown when you log out but the next time you visit for
example /stream, by which time you may well be logged in again.

Note that this create an account behavior is badly broken: You are not logged
in after creating an account, instead it sends you an activation email, but
you aren't shown any message to say that an activation email has been sent.
You're just redirected to the logged-out front page and told nothing. Clicking
on the link in the activation email also just redirects you to the logged-out
front page without logging you in or telling you that you've activated, but it
also sets multiple flash messages ("Please check your email to activate" and
"You have activated" which are all shown at once the next time the Angular app
loads (e.g. when you visit /stream).

This, however, is all the same as the current behaviour in production (!). This
needs to be fixed in a separate commit.