Show flash messages on the home page
This fixes three issues with the new user account registration process:

1. After submitting the form at /register the user is redirected to the
front page, but they aren't shown the flash message notifying them that
we have sent them an activation email.

2. After opening the link in an activation email the user is redirected
to the front page, but they aren't shown the flash message notifying
them that their account has been activated.

3. Instead the user would see both flash messages at the same time the next
time they opened a page containing the stream or sidebar (for example:
after logging in and being redirected to the stream).

This change fixes all three issues by popping and displaying the flash
message on the front page:

1. After submitting the form at /register the user is redirected to the
front page and is shown the flash message notifying them that we have
sent them an activation email.

2. After opening the link in an activation email the user is redirected
to the front page, and is shown the flash message notifying them that
their account has been activated.

3. After logging in the user is redirected to the stream, and they are
not shown any flash message.

This change also fixes the same issue with the reset password process:
after resetting their password the user is redirected to the front page
but is not shown the "Your password has been reset!" flash message, the
message is instead shown to them the next time they visit a page with
the stream or sidebar on it. With this change the message is shown to
them on the front page.