Move render() out of AnnotationController
Move the public method vm.render() out of AnnotationController, making it a
private function and renaming it to updateViewModel().

This function is not part of the view model (it's never called by the
templates, only by code in annotation.js and tests) and its tests don't need to
be part of the annotation controller/directive tests. Move it out.

All the function does is copy some fields from the domain model to the view
model, the opposite of what the updateDomainModel() function does, so rename
it to updateViewModel() and put it next to updateDomainModel(), clarifying its
role.