Move API exception handling into exception views
With a view to making the API view functions easier to read and modify,
this commit moves API exception handling for a number of different error
types into standalone exception views. This eliminates the need for
repetitive boilerplate in view functions, which can simply raise an
exception of the appropriate type.