Remove <base> tag from the main website
The `<base>` tag was set on every page of the main site,
always pointing to '/'. This broke fragment only links on pages
and also, in Firefox and Edge, SVG `<use>` elements whose
'xlink:href' attributes pointed to elements defined via inline
SVGs on the same page.

The `<base>` tag is still needed in the sidebar and stream
by Angular's HTML5 history support (though this could be disabled)
and there are a few places in the app's code that use document.baseURI
in order to construct public links to groups and annotations.

As a start towards removing it everywhere, remove it on the main site
and keep it only in the app.