Don't 500 when non-object JSON is POSTed to API create
Sean just discovered that POSTing data that is valid JSON but not an
object to the annotation create API results in an application crash.

Fix this by simply ignoring anything we don't understand in the
traversal phase and assuming that the annotation to be created is public
unless we can tell otherwise.