Update the truncated state of the annotation when the feature flag changes
If the enabled state of the annotation changes after the view
loaded, the truncation state did not update.

With the current feature flag implementation this typically happens
if the feature flags load after the session data has been received.
In future the intention is to make feature flags part of the session
data payload so this won't happen.