Remove unnecessary pointer-events property
The inline controls use to cover the full height of the right-side
of excerpts, not just the visible area occupied by the 'More' and
'Less' links. This is no longer the case.

The use of 'pointer-events' here also caused the 'More' link
to be unresponsive in Microsoft Edge. There may be a difference
in behavior between Edge and FF/Chrome when nested elements
have different pointer-events values, but I haven't verified
this yet.