Add annotation count to admin user info page
This commit adds a total (public + shared + groups) annotation count to
the user information table displayed in the admin interface. This can be
useful when determining if a user is actively using Hypothesis, or to
correlate with the public API when determining if a user has mistakenly
make private annotations.