Move `uri.expand` to `storage.expand_uri`
Unfortunately, `h.api.storage` currently needs to be able to call
`uri.normalize` (via `h.api.search.transform.prepare`) in order to
normalise URIs in passed annotations. This leads to a circular import
problem because `uri.expand` relies on storage to find equivalent URIs.

Although this isn't particularly satisfactory, the simplest solution
seems to be to move the functionality of `uri.expand` into the storage
module, as it is a thin wrapper over a query of the document model.