Remove circular import between storage and search
Because `h.api.storage` is becoming the central point of indirection for
storage calls, large parts of the application import it. This means that
we have to be very careful what we import from the module.
This commit removes various things which cause the import of
`h.api.storage` from the import path of the `h.api.transform.prepare`.