Add "create account or sign in" banner
Also remove the "Sign in / Create an account" tabs from the "sign in"
panel. They're no longer needed because the new "To annotate this
document create a free account or sign in" banner to the top of the
sidebar has both links.

Also add a "Sign up" link to the top bar.