Show the annotation's group not the focused group
In AnnotationController use domainModel.group instead of
groups.focused() to determine which group to display on the annotation.

groups.focused() works fine as long as the group that the annotation
belongs to is always the currently focused group. This is true when in
the sidebar, but not on the stream or on individual annotation pages. On
those pages the most recently focused group the last time the sidebar
was shown is displayed on every annotation, instead of the actual groups
of the annotations.

This means that we now need to update the domainModel.group of new
annotations when the focused group changes, to support this use:

1. Create a new annotation but don't click save yet
2. Change the focused group
3. Save the annotation

The annotation is correctly saved with the currently (and not the
previously) focused group, and also after the focused group the name of
the newly focused group is displayed on the annotation before it's
saved.

Fixes #2839.