Speed up repeated docker builds
This commit attempts to improve the lives of people frequently building
and deploying the h docker image by increasing the proportion of the
image that remains cached across multiple builds, at the cost of a
somewhat larger image.

Changes include:

- Build dependencies are now installed once and left in the image, as
  opposed to installed only for the duration of the build and then
  removed. The impact of this change is analysed below.
- CHANGES.rst no longer invalidates application dependency build.
- Source files are not chowned to the hypothesis user, so source code
  doesn't take up twice the space in the image.

With this set of changes, at this commit, the total size of the image is
460 MB, compared to 282 MB for the image for v0.8.9. But when we look at
the layer breakdown for each of these images we see a different story:

v0.8.9 image:

    LAYER                                 SIZE        CUMULATIVE
    base image                            5.249 MB    5.249 MB
    system dependencies                   73.88 MB    79.13 MB
    packaging files                       85.55 kB*   79.22 MB
    build and application dependencies    130.5 MB*   209.7 MB
    application source code               64.95 MB*   274.7 MB
    application static asset build        7.571 MB*   282.2 MB

This commit:

    LAYER                                 SIZE        CUMULATIVE
    base image                            5.249 MB    5.249 MB
    build and system dependencies         308.1 MB    313.4 MB
    packaging files                       87.95 kB    313.4 MB
    application dependencies              106.2 MB    419.6 MB
    application source code               30.73 MB*   450.4 MB
    application static asset build        6.145 MB*   456.5 MB

The difference here is that a typical build (one which didn't change the
python or node dependencies) would rebuild each of the starred layers,
amounting to a ~203 MB diff previously, and a ~37 MB diff now. The
difference is smaller but still not insignificant if a build changes the
python/node dependencies (~203 MB/143 MB).

The build is now significantly faster as the system build dependencies
no longer need to be downloaded and installed every time:

                  NO CACHE     WARM CACHE
    v0.8.9        7m 9s        6m 41s
    this commit   7m 15s       57s