Remove an unnecessary $watch()
As far as I can tell this $watch() doesn't do anything. Its purpose
appears to be to uncollapse a thread if that thread contains an
annotation that's being edited. But after deleting it, that uncollapsing
still seems to happen (for example: edit an annotation but don't save or
cancel it yet, change to another group, change back to the first group -
the thread containing the annotation you were editing is uncollapsed).

This code was added before the groups feature existed, so I don't know
what use-case it was intended for.