Build and upload Chrome extensions as build artifacts
Travis supports uploading named files as "build artifacts" after the
tests have run.

This commit adds tasks to the Makefile specifically to build our stage
and production extensions (the extension IDs are public data) and sets
up Travis to upload zipped extension bundles as artifacts.

At the same time it:

a) removes the development builds of the extension from testing to save
   time, as it seems unlikely that we will break the development
   extension and not the production extensions.

b) removes the build of the Firefox extension from testing, as it seems
   very likely that this iteration of the Firefox extension will soon be
   discarded in its entirety in favour of one built on the WebExtensions
   API.