Add WebTrends ?WT.*= params to normalisation blacklist
A URL with a "WT.mc_id" query param on it refers to the same document as
the same URL without the query param. These query params should be
removed from the URL as part of URL normalization.

WebTrends also supports a large number of other query parameters, all of
the form "WT.*" -- these are also excluded.