Add a first-pass Annotation model class and migration
Adds a SQLAlchemy model for storing annotation data. Some notable
changes from the way annotations are currently stored in Elasticsearch:

- The user and group fields are represented internally as `userid` and
  `groupid`, to make clear that they contain the relevant identifiers
  for users and groups, not the objects themselves.
- The `permissions` ACL object has been condensed to a single boolean
  field: `shared`.
- Annotations are associated with only a single target. The URI for the
  target is stored as `target_uri`, with a normalized version in
  `target_uri_normalized`. The serialised selectors for the target are
  stored in a customised JSONB type (to handle some of the strange data
  we get from the TextQuoteSelector) in the `target_selectors` field.
- Data we don't know anything about gets stored in a JSONB field called
  `extra`.

This is very much a first pass at this model. It is likely we will need
to tweak indices on various columns, and perhaps even rethink how
certain information is stored, as we learn how to integrate it with the
rest of the code.