Decouple the Hypothesis service URL from the app's base URI
Previously the URL set as the <base> tag was used indirectly
to specify the URL of the Hypothesis service that various
links referred to.

Make things clearer by adding an explicit 'serviceUrl' attribute
to the Hypothesis client settings and use that in place of the
baseURI.

The custom template whitelist configuration has been removed
because all of the templates are embedded in app.html, so the
default value of ['self'], allowing templates to be loaded
only from the same domain as the page, is fine.