Add script for migrating Elasticsearch annotations to Postgres
Scans annotations out of Elasticsearch in batches of 2000, tries to map
them on to the Postgres Annotation model and then bulk-inserts them into
the database.

When re-running the migration it will start scanning Elasticsearch
annotations where the `updated` field is greater or equal to one hour
earlier to the latest `updated` Postgres annotation.

Things to remember for mapping Elasticsearch annotation to Postgres
ones:

- Maps the ES permissions field to a single `shared` boolean.
- Discards multiple targets in favour of the first one.
- Discards the ES `annotation['uri']` in favour of the
  `annotation['target'][0]['source']`.