Don't open "welcome page" when installed by admin policy
If the extension has been installed as the result of an administrative
policy rather than user action, it doesn't make a lot of sense (and
appears to cause some frustration) to open a tab to our "welcome page."

This commit suppresses the opening of the "welcome page" on first run
if the install type is admin. Chrome documentation references:

- https://developer.chrome.com/extensions/management#event-onInstalled
- https://developer.chrome.com/extensions/management#type-ExtensionInfo
- https://developer.chrome.com/extensions/management#type-ExtensionInstallType

Fixes #2890.