Don't 404 if a user visits an activation link twice
When a user activates their account we delete the row from the activations
database table. If the user then tries to visit an activation link from an
activation email that we sent them, after their account has already been
activated, then we can no longer find the activation row in the db so we 404.

This confuses some users.

Instead, redirect them to the front page (as we do after a successful
activation) with a flash message suggesting that they may already be activated
and should try signing in.

Note that if the user tries to visit a genuinely wrong activation link they
will get this same redirect and suggestion - we don't distinguish between old
activation links and incorrect ones.