Handle activation links when signed in
If the user is already signed in to the account that they're trying to
activate (so presumably the account must already be activated) then
tell them that they're signed in, instead of showing them a message
suggesting that they sign in!

If the user is already signed in two a _different_ account when they
open an activation link, show them a different message.