Do not update the group for existing annotations on edit
Whenever an annotation was edited, its current group
was set to whichever group was focused in the UI.

Our current model does not allow for moving annotations
between groups however. The only time when annotations
should have their groups set is when the annotation
is newly created and when switching groups before saving
the annotation. We already have separate logic to deal
with both of these cases.

Fixes #2902