Don't run honcho in Dockerfile by default
Running multiple processes behind honcho in the Dockerfile is a bit of
an antipattern: it makes it easier to start a complete Hypothesis
instance, perhaps, but at the cost of coupling the provision of worker
processes to the provision of web workers.

This commit updates the Dockerfile and the associated documentation so
that the default container command runs only the web server. Auxiliary
workers (the notification worker, nipsa worker, etc) must be run in
their own containers.