Use --websocket to set client websocket URL in buildext
It's now possible to build the client in such a way that a WebSocket
connection is not opened. Indeed, it's the default if the WebSocket URL
has not been explicitly set as a configuration variable. See 00e3ba5.

This commit adds a `--websocket` option to `hypothesis-buildext`, which
allows setting the WebSocket URL explicitly when building an extension
with a bundled `app.html`. The option has no effect if
`--no-bundle-sidebar` is used.