Ensure admin tests work in all environments
This is ridiculous. These tests are passing on Travis and on other
people's machines, using identical versions of the `mock` and `pytest`
libraries.

Travis is using Python 2.7.9, and another developer is using 2.7.11. In
both those environments,

    mock.MagicMock().count() > 0

is False. On my machine, it's True, which makes sense, as the LHS is an
object.

This commit explicitly sets the expected count of the
`Groups.created_by` helper method to ensure the tests pass on all
environments.

Anyone who works out what was going on here will earn 1 Beverage Of
Their Choice.