Move access_token() into annotator_token()
It turns out that the view that the client actually posts to is
annotator_token(), which just calls access_token(). Nothing posts to
access_token() directly. So just replace annotator_token()'s body with
access_token()'s and delete access_token(). Also simplify the code from
access_token(): it just needs to return the JWT string, not a Response.