Specify the minimum Chrome version in the manifest
We've had several reports of users who were confused by ungraceful
behavior when the extension failed to load under old versions
of Chrome.

The current minimum version that our extension will actually work
on is Chrome 34 but in the interests of defining a simple ongoing
support policy which maps to a clear number of regular releases
and a well-defined time frame, I've rounded that to
'the last 10 stable releases' which makes the minimum version Chrome 38.

We can revise this if necessary.

Card CXkq74c7