Fix handling of 'chrome.*' protocol URLs that contain '.pdf'
Chrome disallows injection of content scripts into chrome-extension://
URLs.

The check for whether the tab's URL was supported was run _after_
the content type was detected, so if the URL contained '.pdf',
the content type would first be detected as a PDF and the extension
would then try to load the viewer which would fail because
it would be unable to fetch the content from the tab's original
chrome-extension:// URL.

This hoists the protocol check logic to happen before trying to
determine the content type.

Fixes #2932