Add a first-pass Document related model class and migration
Adds SQLAlchemy models for Document, DocumentURI, and DocumentMeta.

Instead of just having a Document JSON blob where the list of equivalent
uris is appended to, we try and split the document into three logical
parts.

**Document**
The Document itself that only has an `id` column and timestamps, acts as
the master model to wich both uris and metadata is attached.

**DocumentURI**
A logical document can be accessed by different URIs, all of these URIs
are stored individually as a DocumentURI. Along with other data like a
type and content type.
We use the annotation `target_uri` as the `claimant` and then store each
URI the documents lists as an equivalent URI into the `uri` field, along
with its potential type (e.g. self-claim, or rel attributes) and content type.

**DocumentMeta**
In addition to document uris, we store extra metadata of the document.
There is currently no use of that data yet except for generating
specific DocumentURI models for both dublincore and highwire doi uris.

Similar to the Annotation model, all of the above is subject to change
once we start implementing code that uses these new models.