Improve handling of unexpected errors when injecting sidebar
Clicking on the badge for an errored tab showed a help page,
except for unexpected errors when it did nothing but
threw another error internally.

This makes the behavior consistent by adding an 'other error'
section for the Chrome extension error help page which displays
a brief message and the Error.message property.