Remove versioneer in favour of simpler version management
We were using [versioneer](https://github.com/warner/python-versioneer)
to provide version numbers automatically from source control. It's a
great tool, and has allowed us to mostly ignore manual versioning. We
simply tag a commit and then the built version of h has the right
version.

Unfortunately, due to limitations in git, versioneer can't easily
generate correct version numbers for a `git archive` tarball when not on
a tag commit. The most straightforward way I can see to get correct
version numbers for "off-tag" builds of h is to use the tarball output
by

    python setup.py sdist

when building the Docker image.

But when we impose that restriction, it no longer makes an awful lot of
sense to keep versioneer (and all its many and various ways of
calculating the version) around. We can replicate its functionality for
our simpler and less general setup with a few dozen lines in `setup.py`.