Don't crash out if username blacklist doesn't load
If, for whatever reason, we can't load the username blacklist, then
don't fail to register the user. Instead, log a warning about the
problem and carry on.

This isn't necessarily ideal, as it means that (at least in theory) a
user can register a blacklisted username while this problem exists. But
it's probably better than throwing a 500 and failing to register people.