Use Angular 1.5 one-way bindings where appropriate
Angular 1.5 introduces '<' as a way of specifying
that an isolate scope property is an input (one-way) binding.
Note that this is different from '::' (one-time) bindings from
ng 1.3.

See http://angularjs.blogspot.co.uk/2016/02/angular-150-ennoblement-facilitation.html
for details.

The principle benefit is making it obvious when looking
at the definition of a component whether a property is an input
or whether it is a binding that the component might write to.