Fix embed under IE 10/11
wgxpath is now included in the polyfills.js bundle rather
than as a separate bundle.

Making it a separate bundle which is loaded only as necessary
is actually a sensible thing to do and we could also
extend that to the URL, Promise and Object.assign() polyfills.

We can revisit that after the new build system lands.