Add "bouncer" support to Chrome extension
Add support for "bouncer", the new Hypothesis direct-link bouncer
service, to the Chrome extension.

Some JavaScript the bouncer instance that will be running at
https://hpt.is/ needs to be able to tell whether the user has our Chrome
extension installed or not. This means that our Chrome extension needs
to:

1. Declare in its manifest.json that https://hpt.is/ can send messages
   to it.
2. Listen for and reply to those messages.

Development builds of the Chrome extension will need to accept messages
from dev instances of bouncer, so add an argument to the
hypothesis-buildext command so that developers can specify their dev
instance's local URL instead of https://hpt.is/.