Use the same set of polyfills in Karma tests and the app
As per the comment in karma-phantom-polyfills.js, it is
preferable that the tests use only a minimal set
of polyfills which match those used in the app.

Since we now have a script which requires all of the polyfills
used in the app for IE 10/11, we can just include that
in karma-phantomjs-polyfill.js.

This also removes the need to keep two separate Promise polyfill
packages.