Allow annotations with certain invalid permissons
The client auth state can sometimes be out of sync (see #2924).

This commit tries to fix the data problem when annotations have been
created when the client got confused and put a different user into the
`annotation[user]` and the permission fields.

When we come across an annotation where the permission field values
start with `acct:` but don't match the user of the annotation, we
discard the permission values and use the user of the annotation.